requirements.txt
backoff==1.5.0 # rq.filter: <2.0
lxml==4.3.3
paramiko==2.4.2 # rq.filter: <3.0
pytest==3.5.0 # rq.filter: <4.0
python-dateutil==2.7.2 # rq.filter: <3.0
pytz==2018.4
requests==2.20.0 # rq.filter: <3.0
flake8==3.7.7
